Add files via upload
This worksheet has the final outcome containing all python logic with the final payout statement and charts.  
The end user using the python script code in the prior branches should be able to arrive at the same solution using Python in Excel.  
Let me know how it works for you.